+++ osu.Game/Graphics/UserInterface/StandardComboCounter.cs
@@ -64,9 +64,9 @@ namespace osu.Game.Graphics.UserInterface
             Add(popOutSpriteText);
         }
 
-        protected override ulong getProportionalDuration(ulong currentValue, ulong newValue)
+        protected override double getProportionalDuration(ulong currentValue, ulong newValue)
         {
-            ulong difference = currentValue > newValue ? currentValue - newValue : currentValue - newValue;
+            double difference = currentValue > newValue ? currentValue - newValue : currentValue - newValue;
             return difference * RollingDuration;
         }
